wayland_backend

Module debug

Debugging helpers to handle WAYLAND_DEBUG env variable.

Structs

DisplaySlice 🔒

Functions

has_debug_client_env

The WAYLAND_DEBUG env variable is set to debug client.

print_dispatched_message

Print the dispatched message to stderr in a following format:

print_send_message

Print the send message to stderr in a following format:

print_timestamp 🔒

Print timestamp in seconds.microseconds format.