Struct wayland_backend::rs::client::Handle

pub struct Handle { /* fields omitted */ }

Main handle of a backend to the Wayland protocol

This type hosts most of the protocol-related functionality of the backend, and is the main entry point for manipulating Wayland objects. It can be retrieved both from the backend via Backend::handle(), and is given to you as argument in most event callbacks.

Implementations

impl Handle

pub fn display_id(&self) -> ObjectId

Get the object ID for the wl_display

pub fn last_error(&self) -> Option<WaylandError>

Get the last error that occurred on this backend

If this returns an error, your Wayland connection is already dead.

pub fn info(&self, id: ObjectId) -> Result<ObjectInfo, InvalidId>

Get the detailed information about a wayland object

Returns an error if the provided object ID is no longer valid.

pub fn null_id(&mut self) -> ObjectId

Create a null object ID

This object ID is always invalid, and can be used as placeholder.

pub fn placeholder_id(
&mut self,
spec: Option<(&'static Interface, u32)>
) -> ObjectId

Create a placeholder ID for object creation

This ID needs to be created beforehand and given as argument to a request creating a new object ID. A specification must be specified if the interface and version cannot be inferred from the protocol (for example object creation from the wl_registry).

If a specification is provided it’ll be checked against what can be deduced from the protocol specification, and send_request will panic if they do not match.

pub fn send_request(
&mut self,
_: Message<ObjectId>,
data: Option<Arc<dyn ObjectData>>
) -> Result<ObjectId, InvalidId>

Sends a request to the server

Returns an error if the sender ID of the provided message is no longer valid.

Panic:

Several checks against the protocol specification are done, and this method will panic if they do not pass:

the message opcode must be valid for the sender interface

the argument list must match the prototype for the message associated with this opcode

if the method creates a new object, a placeholder_id() must be given in the argument list, either without a specification, or with a specification that matches the interface and version deduced from the protocol rules

pub fn get_data(&self, id: ObjectId) -> Result<Arc<dyn ObjectData>, InvalidId>

Access the object data associated with a given object ID

Returns an error if the object ID is not longer valid

pub fn set_data(
&mut self,
id: ObjectId,
data: Arc<dyn ObjectData>
) -> Result<(), InvalidId>

Set the object data associated with a given object ID

Returns an error if the object ID is not longer valid