wayland_backend::rs::client

Struct ObjectId

pub struct ObjectId { /* private fields */ }

An ID representing a Wayland object

The backend internally tracks which IDs are still valid, invalidates them when the protocol object they represent is destroyed. As such even though the Wayland protocol reuses IDs, you can confidently compare two ObjectId for equality, they will only compare as equal if they both represent the same protocol object.

Implementations

impl ObjectId

pub fn is_null(&self) -> bool

Check if this is a null ID

Note: This is not the same as checking if the ID is still valid, which cannot be done without the Backend. A null ID is the ID equivalent of a null pointer: it never has been valid and never will be.

pub fn null() -> ObjectId

Create a null object ID

This object ID is always invalid, and should be used as placeholder in requests that create objects, or for request with an optional Object argument.

See Backend::send_request() for details.

pub fn interface(&self) -> &'static Interface

Interface of the represented object

pub fn protocol_id(&self) -> u32

Return the protocol-level numerical ID of this object

Protocol IDs are reused after object destruction, so this should not be used as a unique identifier, instead use the ObjectId directly, it implements Clone, PartialEq, Eq and Hash.