Module wayland_backend::rs::server

Server-side rust implementation of a Wayland protocol backend

The main entrypoint is the Backend::new method.

Structs

Backend

A backend object that represents the state of a wayland server.

ClientId

An ID representing a Wayland client

Credentials

Holds the client credentials

GlobalId

An Id representing a global

GlobalInfo

Description of a global advertised to some clients.

Handle

Main handle of a backend to the Wayland protocol

InvalidId

An error generated when trying to act on an invalid ObjectId.

ObjectId

An ID representing a Wayland object

WeakHandle

A weak reference to a Handle

Enums

DisconnectReason

Describes why a client has been disconnected from the server.

InitError

An error type representing the failure to initialize a backend

Traits

ClientData

A trait representing your data associated to a client

GlobalHandler

A trait representing the handling of new bound globals

ObjectData

A trait representing your data associated to an object