Trait wayland_backend::rs::server::ClientData

pub trait ClientData: DowncastSync {
    // Provided methods
    fn initialized(&self, _client_id: ClientId) { ... }
    fn disconnected(&self, _client_id: ClientId, _reason: DisconnectReason) { ... }
    fn debug(&self, f: &mut Formatter<'_>) -> Result { ... }
}

A trait representing your data associated to a client

Provided Methods

fn initialized(&self, _client_id: ClientId)

Notification that the client was initialized

fn disconnected(&self, _client_id: ClientId, _reason: DisconnectReason)

Notification that the client is disconnected

fn debug(&self, f: &mut Formatter<'_>) -> Result

Helper for forwarding a Debug implementation of your ClientData type

By default will just print GlobalHandler { ... }

Implementations

impl dyn ClientData

pub fn is<__T: ClientData>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: ClientData>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: ClientData>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: ClientData>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: ClientData>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_arc<__T>(self: Arc<Self>) -> Result<Arc<__T>, Arc<Self>>
where __T: Any + Send + Sync + ClientData,

Returns an Arc-ed object from an Arc-ed trait object if the underlying object is of type __T. Returns the original Arc-ed trait if it isn’t.