wayland_backend::rs::server

Trait ObjectData

pub trait ObjectData<D>: DowncastSync {
    // Required methods
    fn request(
        self: Arc<Self>,
        handle: &Handle,
        data: &mut D,
        client_id: ClientId,
        msg: Message<ObjectId, OwnedFd>,
    ) -> Option<Arc<dyn ObjectData<D>>>;
    fn destroyed(
        self: Arc<Self>,
        handle: &Handle,
        data: &mut D,
        client_id: ClientId,
        object_id: ObjectId,
    );

    // Provided method
    fn debug(&self, f: &mut Formatter<'_>) -> Result { ... }
}

A trait representing your data associated to an object

You will only be given access to it as a & reference, so you need to handle interior mutability by yourself.

The methods of this trait will be invoked internally every time a new object is created to initialize its data.

Required Methods

fn request( self: Arc<Self>, handle: &Handle, data: &mut D, client_id: ClientId, msg: Message<ObjectId, OwnedFd>, ) -> Option<Arc<dyn ObjectData<D>>>

Dispatch a request for the associated object

If the request has a NewId argument, the callback must return the object data for the newly created object

fn destroyed( self: Arc<Self>, handle: &Handle, data: &mut D, client_id: ClientId, object_id: ObjectId, )

Notification that the object has been destroyed and is no longer active

Provided Methods

fn debug(&self, f: &mut Formatter<'_>) -> Result

Helper for forwarding a Debug implementation of your ObjectData type

By default will just print ObjectData { ... }

Implementations

impl<D> dyn ObjectData<D>
where D: Any + 'static,

pub fn is<__T: ObjectData<D>>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: ObjectData<D>>( self: Box<Self>, ) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: ObjectData<D>>( self: Rc<Self>, ) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: ObjectData<D>>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: ObjectData<D>>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_arc<__T: ObjectData<D> + Any + Send + Sync>( self: Arc<Self>, ) -> Result<Arc<__T>, Arc<Self>>

Returns an Arc-ed object from an Arc-ed trait object if the underlying object is of type __T. Returns the original Arc-ed trait if it isn’t.