wayland_backend::rs

Module wire

Types and routines used to manipulate arguments from the wire format

Enums

MessageParseError

Error generated when trying to deserialize a message from buffers

MessageWriteError

Error generated when trying to serialize a message into buffers

Functions

next_multiple_of 🔒

parse_message

Attempts to parse a single wayland message with the given signature.

write_to_buffers

Serialize the contents of this message into provided buffers