wayland_backend

Module sys

Available on crate features client_system or server_system only.

Implementations of the Wayland backends using the system libwayland

Modules

clientclient_system

Client-side implementation of a Wayland protocol backend using libwayland

serverserver_system

Server-side implementation of a Wayland protocol backend using libwayland