Trait wayland_backend::sys::server::ClientData

pub trait ClientData<D>: DowncastSync {
    fn initialized(&self, client_id: ClientId);

    fn disconnected(&self, client_id: ClientId, reason: DisconnectReason);

    fn debug(&self, f: &mut Formatter<'_>) -> Result { ... }
}

A trait representing your data associated to a clientObjectData

Required methods

fn initialized(&self, client_id: ClientId)

Notification that a client was initialized

fn disconnected(&self, client_id: ClientId, reason: DisconnectReason)

Notification that a client is disconnected

Provided methods

fn debug(&self, f: &mut Formatter<'_>) -> Result

Helper for forwarding a Debug implementation of your ClientData type

By default will just print GlobalHandler { ... }

Implementations

impl<D> dyn ClientData<D> where
D: Any + 'static,

pub fn is<__T: ClientData<D>>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: ClientData<D>>(
self: Box<Self>
) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: ClientData<D>>(
self: Rc<Self>
) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: ClientData<D>>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: ClientData<D>>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_arc<__T: ClientData<D>>(
self: Arc<Self>
) -> Result<Arc<__T>, Arc<Self>> where
__T: Any + Send + Sync,

Returns an Arc-ed object from an Arc-ed trait object if the underlying object is of type __T. Returns the original Arc-ed trait if it isn’t.