wayland_backend::types

Module server

Structs

Credentials

Holds the client credentials

GlobalInfo

Description of a global advertised to some clients.

InvalidId

An error generated when trying to act on an invalid ObjectId.

Enums

DisconnectReason

Describes why a client has been disconnected from the server.

InitError

An error type representing the failure to initialize a backend