Module wayland_backend::protocol

Types and utilities for manipulating the Wayland protocol

Structs

Interface

Description of wayland interface.

Message

Represents a message that has been sent from some object.

MessageDesc

Wire metadata of a given message

ObjectInfo

Description of the protocol-level information of an object

ProtocolError

A protocol error

WEnumError

Error representing an unknown numeric variant for a WEnum

wl_interface

wl_message

Enums

AllowNull

Describes whether an argument may have a null value.

Argument

Enum of possible argument of the protocol

ArgumentType

Enum of possible argument types as recognized by the wire

WEnum

An enum value in the protocol.

Constants

INLINE_ARGS

Number of arguments that are stocked inline in a Message before allocating

Statics

ANONYMOUS_INTERFACE

Special interface representing an anonymous object

Functions

same_interface

Returns true if the two interfaces are the same.

Unions

wl_argument