Module wayland_backend::rs::client

Client-side rust implementation of a Wayland protocol backend

Structs

Backend

A pure rust implementation of a Wayland client backend

Handle

Main handle of a backend to the Wayland protocol

InvalidId

An error generated when trying to act on an invalid ObjectId.

NoWaylandLib

An error type representing the failure to load libwayland

ObjectId

An ID representing a Wayland object

ReadEventsGuard

Guard for synchronizing event reading across multiple threads

Enums

WaylandError

An error that can occur when using a Wayland connection

Traits

ObjectData

A trait representing your data associated to an object