Module wayland_backend::sys::client

Client-side implementation of a Wayland protocol backend using libwayland

Entrypoints are:

Backend::connect method if you’re creating the Wayland connection

Backend::from_foreign_display if you’re interacting with an already existing Wayland connection through FFI.

Structs

Backend

A Wayland client backend

InvalidId

An error generated when trying to act on an invalid ObjectId.

NoWaylandLib

An error type representing the failure to load libwayland

ObjectId

An ID representing a Wayland object

ReadEventsGuard

Guard for synchronizing event reading across multiple threads

WeakBackend

A weak handle to a Backend

Enums

WaylandError

An error that can occur when using a Wayland connection

Traits

ObjectData

A trait representing your data associated to an object