Struct wayland_backend::sys::client_impl::InnerBackend

pub struct InnerBackend {
    inner: Arc<Inner>,
}

Fields
inner: Arc<Inner>
Implementations

impl InnerBackend

fn lock_state(&self) -> MutexGuard<'_, ConnectionState>

pub fn downgrade(&self) -> WeakInnerBackend

pub fn display_ptr(&self) -> *mut wl_display

impl InnerBackend

pub fn connect(stream: UnixStream) -> Result<Self, NoWaylandLib>

pub unsafe fn from_foreign_display(display: *mut wl_display) -> Self

fn from_display(display: *mut wl_display, owned: bool) -> Self

pub fn flush(&self) -> Result<(), WaylandError>

pub fn poll_fd(&self) -> BorrowedFd<'_>

pub fn dispatch_inner_queue(&self) -> Result<usize, WaylandError>

impl InnerBackend

pub fn display_id(&self) -> ObjectId

pub fn last_error(&self) -> Option<WaylandError>

pub fn info(&self, _: ObjectId) -> Result<ObjectInfo, InvalidId>

pub fn null_id() -> ObjectId

pub fn send_request( &self, _: Message<ObjectId, RawFd>, data: Option<Arc<dyn ObjectData>>, child_spec: Option<(&'static Interface, u32)>, ) -> Result<ObjectId, InvalidId>

pub fn get_data(&self, _: ObjectId) -> Result<Arc<dyn ObjectData>, InvalidId>

pub fn set_data( &self, _: ObjectId, data: Arc<dyn ObjectData>, ) -> Result<(), InvalidId>

pub unsafe fn manage_object( &self, interface: &'static Interface, proxy: *mut wl_proxy, data: Arc<dyn ObjectData>, ) -> ObjectId

Start managing a Wayland object.

Safety: This will change the event queue the proxy is associated with. Changing the event queue of an existing proxy is not thread-safe. If another thread is concurrently reading the wayland socket and the proxy already received an event it might get enqueued on the old event queue.

unsafe fn manage_object_internal( &self, interface: &'static Interface, proxy: *mut wl_proxy, data: Arc<dyn ObjectData>, guard: &mut MutexGuard<'_, ConnectionState>, ) -> ObjectId

Start managing a Wayland object.

Opposed to Self::manage_object, this does not acquire any guards.